+++ include/H2DE/surfaces/H2DE_surface.h
@@ -234,5 +234,6 @@ private:
 #include <H2DE/surfaces/H2DE_texture.h>
 #include <H2DE/surfaces/H2DE_sprite.h>
 #include <H2DE/surfaces/H2DE_color.h>
+#include <H2DE/surfaces/H2DE_border.h>
 
 #endif