+++ src/surfaces/H2DE_border.cpp
@@ -1,0 +1,20 @@
+#include "H2DE/surfaces/H2DE_border.h"
+#include "H2DE/engine/H2DE_lerp_manager.h"
+
+// INIT
+H2DE_Border::H2DE_Border(H2DE_Engine* e, H2DE_Object* o, const H2DE_SurfaceData& sd, const H2DE_BorderData& bd) noexcept : H2DE_Surface(e, o, sd), borderData(bd) {
+
+}
+
+// SETTER
+H2DE_TimelineID H2DE_Border::setColor(const H2DE_ColorRGB& color, H2DE_TimelineID duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive) {
+    return H2DE_LerpManager::lerp(engine, borderData.color, color, duration, easing, [this](H2DE_ColorRGB iv) {
+        setColor(iv);
+    }, completed, pauseSensitive);
+}
+
+H2DE_TimelineID H2DE_Border::setThickness(uint16_t thickness, H2DE_TimelineID duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive) {
+    return H2DE_LerpManager::lerp<uint16_t>(engine, borderData.thickness, thickness, duration, easing, [this](uint16_t iv) {
+        setThickness(iv);
+    }, completed, pauseSensitive);
+}